+++ frontend/src/components/molecules/ChartCards/StableCoinSupplyBarChart.vue
@@ -4,18 +4,18 @@
 			:stable-coins-data="stableCoinsData"
 			chart-type="supply"
 			show-sign=""
-			:label-clickable="false"
+			:label-clickable="true"
 		>
 			<template #title> Protocol Token Supply </template>
 		</DistributionChartBar>
 	</div>
 </template>
 <script lang="ts" setup>
-import type { StablecoinResponse } from '~/queries/useStableCoinQuery'
+import type { Plttoken } from '~/types/generated'
 import DistributionChartBar from '../DistributionChartBar.vue'
 
 type Props = {
-	stableCoinsData?: StablecoinResponse
+	stableCoinsData?: Plttoken[]
 	isLoading?: boolean
 }