+++ backend/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Changed
+- Aligned with Rust SDK version `7.0.0-alpha.3`.
+
+### Added 
+- Added plt specific tables `plt_events`, `plt_account`, `plt_tokens` 
+- Added query `PltEvents` to fetch events related to plt tokens. 
+- Added query `PltToken` to fetch details of a specific plt token.
 
 ## [2.0.14] - 2025-07-10